+++ components/World/LODController.tsx
@@ -8,6 +8,7 @@
 
 import React, { useEffect } from 'react';
 import { useThree } from '@react-three/fiber';
+import * as THREE from 'three';
 import { getPerformanceManager, QualityLevel } from '../../infrastructure/performance/PerformanceManager';
 import { getDynamicCullingManager } from '../../infrastructure/rendering/DynamicCullingManager';